+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
+## [Unreleased]
+
+### Fixed
+- Check for offline refresh token grant based on refresh token request parameter
+
 ## [22.0.10-1.5] - 2024-06-13
 
 ### Added